+++ ice-20201109/alibabacloud_ice20201109/client.py
@@ -941,6 +941,50 @@ class Client(OpenApiClient):
         runtime = util_models.RuntimeOptions()
         return await self.get_template_with_options_async(request, runtime)
 
+    def get_template_materials_with_options(
+        self,
+        request: ice20201109_models.GetTemplateMaterialsRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> ice20201109_models.GetTemplateMaterialsResponse:
+        UtilClient.validate_model(request)
+        query = OpenApiUtilClient.query(UtilClient.to_map(request))
+        req = open_api_models.OpenApiRequest(
+            query=query
+        )
+        return TeaCore.from_map(
+            ice20201109_models.GetTemplateMaterialsResponse(),
+            self.do_rpcrequest('GetTemplateMaterials', '2020-11-09', 'HTTPS', 'GET', 'AK', 'json', req, runtime)
+        )
+
+    async def get_template_materials_with_options_async(
+        self,
+        request: ice20201109_models.GetTemplateMaterialsRequest,
+        runtime: util_models.RuntimeOptions,
+    ) -> ice20201109_models.GetTemplateMaterialsResponse:
+        UtilClient.validate_model(request)
+        query = OpenApiUtilClient.query(UtilClient.to_map(request))
+        req = open_api_models.OpenApiRequest(
+            query=query
+        )
+        return TeaCore.from_map(
+            ice20201109_models.GetTemplateMaterialsResponse(),
+            await self.do_rpcrequest_async('GetTemplateMaterials', '2020-11-09', 'HTTPS', 'GET', 'AK', 'json', req, runtime)
+        )
+
+    def get_template_materials(
+        self,
+        request: ice20201109_models.GetTemplateMaterialsRequest,
+    ) -> ice20201109_models.GetTemplateMaterialsResponse:
+        runtime = util_models.RuntimeOptions()
+        return self.get_template_materials_with_options(request, runtime)
+
+    async def get_template_materials_async(
+        self,
+        request: ice20201109_models.GetTemplateMaterialsRequest,
+    ) -> ice20201109_models.GetTemplateMaterialsResponse:
+        runtime = util_models.RuntimeOptions()
+        return await self.get_template_materials_with_options_async(request, runtime)
+
     def list_all_public_media_tags_with_options(
         self,
         request: ice20201109_models.ListAllPublicMediaTagsRequest,